Programación en tres capas con módulos correspondientes

+++ Entidades/Articulo.cs
@@ -12,13 +12,13 @@ namespace Entidades
         public string Nombre { get; set; }
         public Categoria Categoria { get; set; }
         public Proveedor Proveedor { get; set; }
-        public float PrecioVenta { get; set; }
-        public float PrecioCompra { get; set; }
+        public decimal PrecioVenta { get; set; }
+        public decimal PrecioCompra { get; set; }
         public int Stock { get; set; }
         public bool Estado { get; set; }
         public Articulo() { }
 
-        public Articulo(int id, string nombre, Categoria categoria, Proveedor proveedor, float precioVenta, float precioCompra, int stock, bool estado)
+        public Articulo(int id, string nombre, Categoria categoria, Proveedor proveedor, decimal precioVenta, decimal precioCompra, int stock, bool estado)
         {
             Id = id;
             Nombre = nombre;
